set up all the pages for JP and ZH
