+++ src/components/logic/translateLink.ts
@@ -1,0 +1,9 @@
+import * as language from "../../data/languages"
+
+export default (link, currentLanguage) => {
+  if (currentLanguage === language.defaultLangKey) {
+    return link
+  }
+
+  return `${currentLanguage}/${link}`
+}
