+++ src/data/generic.tsx
@@ -1,0 +1,7 @@
+export default {
+  copy: {
+    en: '',
+    jp: '',
+    zh: '',
+  }
+}
