+++ src/data/nav.tsx
@@ -13,8 +13,20 @@ export default {
     faqs: "FAQs",
     releases: "Releases",
   },
+  jp: {
+    home: "",
+    getStarted: "Get Started",
+    advanced: "Advanced",
+    builder: (
+      <>
+        <span className="desktopOnly">Form&nbsp;</span>Builder
+      </>
+    ),
+    faqs: "FAQs",
+    releases: "Releases",
+  },
   zh: {
-    home: "Home",
+    home: "首页",
     getStarted: "Get Started",
     advanced: "Advanced",
     builder: (
