+++ src/pages/faq.jp.tsx
