+++ src/pages/index.jp.tsx
@@ -1,0 +1,25 @@
+import * as React from "react"
+import Layout from "../components/layout"
+import Seo from "../components/seo"
+import Home from "../components/Home"
+import home from "../data/home";
+
+const IndexPage = ({
+  location,
+}: {
+  location: {
+    search: string
+    pathname: string
+  }
+}) => (
+  <Layout location={location}>
+    <Seo
+      title={home.title.en}
+      description={home.description.en}
+      location={location}
+    />
+    <Home location={location} />
+  </Layout>
+)
+
+export default IndexPage
